move allowed functions to utils.js
